monofloor-admin/requirements.txt
python-docx==1.1.0
Pillow==10.1.0
